+++ NEWS.md
@@ -8,6 +8,8 @@
 - Ajout fonction ajouter_typologie() qui permet de créer des agrégats par typologie de territoire et type de territoire (epci, départements, région...)
 - Ajout "arn_plm", table des arrondissements de Paris, Lyon et Marseille
 - Ajout  "table_passage_communes_zonages", table de passage des communes vers des zonages supra
+- Ajout fonction charger_zonage() qui permet de créer un zonage supra-communal adapté aux fonctions ajouter_zonage() et ajouter_typologie()
+
 
 # COGiter 0.0.7